# BUILD/SOURCES/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "AssetMarketPlace"
    },
    "paths": {
...
        },"/getAssetFile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "parameters": [
                    {
                        "name": "caseRef",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "type": "string",
                        "format": ""
                    },
                    {
                        "name": "sandBox",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "type": "string",
                        "format": ""
                    },
                    {
                        "name": "fileName",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "type": "string",
                        "format": ""
                    }
                ],
                "operationId": "getGetAssetFile"
            }
        }
    },
...
}